ci: migrate DLLM tests to test/registered/dllm/
- Move test_llada2_mini.py to test/registered/dllm/ (CUDA, 520s)
- Move test_llada2_mini_amd.py to test/registered/dllm/ (AMD, 520s)
- Add CI registration decorators for both test files
- Remove entries from legacy test/srt/run_suite.py

Part of #13808
