+++ test/registered/dllm/test_llada2_mini.py
@@ -1,8 +1,7 @@
-"""
-Test LLaDA2 (Diffusion Language Model) on AMD GPUs.
+from sglang.test.ci.ci_register import register_cuda_ci
 
-This test verifies that DLLM works on AMD with triton attention backend.
-"""
+# DLLM LLaDA2 tests (1-GPU)
+register_cuda_ci(est_time=520, suite="stage-b-test-small-1-gpu")
 
 import unittest
 from types import SimpleNamespace
@@ -14,13 +13,14 @@ from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
     DEFAULT_URL_FOR_TEST,
     CustomTestCase,
+    is_in_amd_ci,
     is_in_ci,
     popen_launch_server,
     write_github_step_summary,
 )
 
 
-class TestLLaDA2MiniAMD(CustomTestCase):
+class TestLLaDA2Mini(CustomTestCase):
     @classmethod
     def setUpClass(cls):
         cls.model = "inclusionAI/LLaDA2.0-mini"
@@ -33,9 +33,9 @@ class TestLLaDA2MiniAMD(CustomTestCase):
             "--max-running-requests",
             "1",
             "--attention-backend",
-            "triton",  # Use triton for AMD instead of flashinfer
+            "flashinfer",
             "--dllm-algorithm",
-            "LowConfidence",
+            "LowConfidence",  # TODO: Add dLLM configurations
         ]
 
         cls.process = popen_launch_server(
@@ -50,7 +50,6 @@ class TestLLaDA2MiniAMD(CustomTestCase):
         kill_process_tree(cls.process.pid)
 
     def test_gsm8k(self):
-        """Test GSM8K accuracy with DLLM on AMD."""
         args = SimpleNamespace(
             num_shots=5,
             data_path=None,
@@ -63,12 +62,10 @@ class TestLLaDA2MiniAMD(CustomTestCase):
         metrics = run_eval_few_shot_gsm8k(args)
         print(f"{metrics=}")
 
-        # Relaxed thresholds for AMD - may need adjustment
-        self.assertGreater(metrics["accuracy"], 0.80)
-        self.assertGreater(metrics["output_throughput"], 50)
+        self.assertGreater(metrics["accuracy"], 0.88)
+        self.assertGreater(metrics["output_throughput"], 150)
 
     def test_bs_1_speed(self):
-        """Test single batch inference speed."""
         args = BenchArgs(port=int(self.base_url.split(":")[-1]), max_new_tokens=2048)
         acc_length, speed = send_one_prompt(args)
 
@@ -76,11 +73,13 @@ class TestLLaDA2MiniAMD(CustomTestCase):
 
         if is_in_ci():
             write_github_step_summary(
-                f"### test_bs_1_speed (llada2-mini AMD) with tp1\n"
+                f"### test_bs_1_speed (llada2-mini) with tp1\n"
                 f"{speed=:.2f} token/s\n"
             )
-            # Relaxed threshold for AMD
-            self.assertGreater(speed, 10)
+            if is_in_amd_ci():
+                self.assertGreater(speed, 10)
+            else:
+                self.assertGreater(speed, 250)
 
 
 if __name__ == "__main__":
