+++ test/registered/dllm/test_llada2_mini_amd.py
@@ -1,3 +1,14 @@
+from sglang.test.ci.ci_register import register_amd_ci
+
+# DLLM LLaDA2 tests for AMD (1-GPU)
+register_amd_ci(est_time=520, suite="stage-b-test-small-1-gpu")
+
+"""
+Test LLaDA2 (Diffusion Language Model) on AMD GPUs.
+
+This test verifies that DLLM works on AMD with triton attention backend.
+"""
+
 import unittest
 from types import SimpleNamespace
 
@@ -8,14 +19,13 @@ from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
     DEFAULT_URL_FOR_TEST,
     CustomTestCase,
-    is_in_amd_ci,
     is_in_ci,
     popen_launch_server,
     write_github_step_summary,
 )
 
 
-class TestLLaDA2Mini(CustomTestCase):
+class TestLLaDA2MiniAMD(CustomTestCase):
     @classmethod
     def setUpClass(cls):
         cls.model = "inclusionAI/LLaDA2.0-mini"
@@ -28,9 +38,9 @@ class TestLLaDA2Mini(CustomTestCase):
             "--max-running-requests",
             "1",
             "--attention-backend",
-            "flashinfer",
+            "triton",  # Use triton for AMD instead of flashinfer
             "--dllm-algorithm",
-            "LowConfidence",  # TODO: Add dLLM configurations
+            "LowConfidence",
         ]
 
         cls.process = popen_launch_server(
@@ -45,6 +55,7 @@ class TestLLaDA2Mini(CustomTestCase):
         kill_process_tree(cls.process.pid)
 
     def test_gsm8k(self):
+        """Test GSM8K accuracy with DLLM on AMD."""
         args = SimpleNamespace(
             num_shots=5,
             data_path=None,
@@ -57,10 +68,12 @@ class TestLLaDA2Mini(CustomTestCase):
         metrics = run_eval_few_shot_gsm8k(args)
         print(f"{metrics=}")
 
-        self.assertGreater(metrics["accuracy"], 0.88)
-        self.assertGreater(metrics["output_throughput"], 150)
+        # Relaxed thresholds for AMD - may need adjustment
+        self.assertGreater(metrics["accuracy"], 0.80)
+        self.assertGreater(metrics["output_throughput"], 50)
 
     def test_bs_1_speed(self):
+        """Test single batch inference speed."""
         args = BenchArgs(port=int(self.base_url.split(":")[-1]), max_new_tokens=2048)
         acc_length, speed = send_one_prompt(args)
 
@@ -68,13 +81,11 @@ class TestLLaDA2Mini(CustomTestCase):
 
         if is_in_ci():
             write_github_step_summary(
-                f"### test_bs_1_speed (llada2-mini) with tp1\n"
+                f"### test_bs_1_speed (llada2-mini AMD) with tp1\n"
                 f"{speed=:.2f} token/s\n"
             )
-            if is_in_amd_ci():
-                self.assertGreater(speed, 10)
-            else:
-                self.assertGreater(speed, 250)
+            # Relaxed threshold for AMD
+            self.assertGreater(speed, 10)
 
 
 if __name__ == "__main__":
